event-management-system/backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.1
django-cors-headers==4.4.0
dj-database-url==2.2.0
psycopg[binary]==3.2.1
python-dotenv==1.0.1
pytest==8.3.2
pytest-django==4.8.0
